feat(admin): refactor notifications with tabs
