+++ apps/admin/package.json
@@ -36,7 +36,8 @@
     "react-dom": "^18.2.0",
     "react-easy-crop": "^5.5.0",
     "react-router-dom": "^6.23.1",
-    "zod": "^3.24.1"
+    "zod": "^3.24.1",
+    "@radix-ui/react-tabs": "^1.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
